dashboard_service/requirements.txt
django>=4.0
channels
channels-redis
fastapi